Move logger from server to core
