+++ packages/server/src/index.ts
@@ -1,8 +1,9 @@
+import { parseLogLevel } from '@medplum/core';
 import express from 'express';
+import gracefulShutdown from 'http-graceful-shutdown';
 import { initApp, shutdownApp } from './app';
 import { loadConfig } from './config';
-import { globalLogger, parseLogLevel } from './logger';
-import gracefulShutdown from 'http-graceful-shutdown';
+import { globalLogger } from './logger';
 
 export async function main(configName: string): Promise<void> {
   process.on('unhandledRejection', (err: any) => {
